Split ycm logic out of proxy.py
